refactor(imports): migrate remaining deep imports to barrel
Consolidate 3 resolver spec/schemas deep imports to models
barrel: dreamcatcher.ts, wolfQueen.ts, piper.ts now import
SCHEMAS from models barrel instead of models/roles/spec/schemas.

All source-file deep imports within game-engine now use the
closest available barrel (models, models/roles/spec, types).
