+++ packages/game-engine/src/resolvers/dreamcatcher.ts
@@ -7,7 +7,7 @@
  * NOTE: Nightmare block guard is handled at actionHandler layer (single-point guard).
  */
 
-import { SCHEMAS } from '../models/roles/spec/schemas';
+import { SCHEMAS } from '../models';
 import { validateConstraints } from './constraintValidator';
 import type { ResolverFn } from './types';
