+++ packages/game-engine/src/resolvers/piper.ts
@@ -10,7 +10,7 @@
  * NOTE: Nightmare block guard is handled at actionHandler layer (single-point guard).
  */
 
-import { SCHEMAS } from '../models/roles/spec/schemas';
+import { SCHEMAS } from '../models';
 import { validateConstraints } from './constraintValidator';
 import type { ResolverContext, ResolverFn } from './types';
